split index.ts into data.ts and form.ts
